Frontend ready for backend integration, currently implementing backend
Set up the async fetch for logging a workout which consists of a workout with multiple exercises in it. Added all the SQLAlchemy tables but still figuring out the setup, also need to get auth tokens working so I can track whether a user is logged in and then adding the workout to that user.