Edited dashboard to have expandable information and further mini changes
The dashboard will show lots of different information, one section will be previous recent workouts which can be expanded to show the exercises and their specifics within them. Also made it so after logging in the user is taken directly to the dashboard, but actually that should be the workout page so I'll change that. Added a todo to track what I want easier.